moved classes from base to mysql
